hardware file is finished, except for gated mode. At the current project, gated mode is not needed. people who would need gated mode,shall proceed with the coding.

+++ hardware/picoquant/hydraharp400/hharp400.py
@@ -294,7 +294,7 @@ class HydraHarp400(Base, FastCounterInterface):
         else:
             self.tryfunc(self.dll.HH_GetHistogram(self._deviceID, c_counts, 1, 0), "GetHistogram")
 
-        time_trace = py_counts
+        time_trace = np.int64(py_counts)
 
         meas_t = int(self.get_measurement_time())
         info_dict = {'elapsed_sweeps': None,
